requirements.txt
selenium
bs4
requests
pillow
replicate
langchain
langchain_google_genai
pydantic
numpy
scipy
moviepy
nltk
soundfile
beautifulsoup4
langchain-core
pyttsx3
streamlit
coqui-tts
torch
langchain-google-genai
python-dotenv
nest_asyncio
pandas